Split quinn module into endpoint and connection parts
